+++ src/zope/interface/_zope_interface_coptimizations.c
@@ -330,6 +330,7 @@ SB_extends(SB* self, PyObject* other)
 
     implied = self->_implied;
     if (implied == NULL) {
+        PyErr_SetString(PyExc_AttributeError, "_implied");
         return NULL;
     }
 
@@ -619,8 +620,10 @@ CPB_descr_get(CPB* self, PyObject* inst, PyObject* cls)
 {
     PyObject* implements;
 
-    if (self->_cls == NULL)
+    if (self->_cls == NULL) {
+        PyErr_SetString(PyExc_AttributeError, "_cls");
         return NULL;
+    }
 
     if (cls == self->_cls) {
         if (inst == NULL) {
@@ -792,6 +795,7 @@ IB__adapt__(PyObject* self, PyObject* obj)
 
         implied = ((SB*)decl)->_implied;
         if (implied == NULL) {
+            PyErr_SetString(PyExc_AttributeError, "_implied");
             Py_DECREF(decl);
             return NULL;
         }